Extracted connectedCalendar logic for reuse
